LayoutTests/platform/glib/http/tests/websocket/tests/hybi/broken-utf8-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/broken-utf8' failed: Received invalid WebSocket data from the server
Test whether WebSocket aborts the connection when it receives a text frame containing broken UTF-8 data.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
